+++ lib/unitwise/ext.rb
@@ -1,2 +1,5 @@
 require 'unitwise'
 require 'unitwise/ext/numeric'
+
+warn '`unitwise/ext` is deprecated. These core extensions will be removed ' \
+     'in a future version. Please use `unitwise` instead.'