+++ lib/unitwise/ext/numeric.rb
@@ -1,6 +1,8 @@
 # Unitwise extends Numeric to add these dyanmic method conveniences: `1.meter`,
-# `26.2.to_mile`, and `4.convert_to("Joule")`. These overrides are optional due
-# to their controversial nature. `require 'unitwise/ext'` to enable them.
+# `26.2.to_mile`, and `4.convert_to("Joule")`. These overrides are optional
+# and may be enabled with `require 'unitwise/ext'`. These methods are known to
+# reduce performance, as well as violate Ruby best practices. They are
+# deprecated and will be removed in a future version.
 class Numeric
   # Converts numeric to a measurement
   # @param unit [Unitwise::Unit, String] The unit to use in the measurement